General refactoring of main components across the application. Seperated component logic in designated folders/files. Added some type declarations where needed.
