+++ ui/package.json
@@ -30,6 +30,7 @@
   },
   "scripts": {
     "start": "craco start",
+    "dev": "nodemon craco start",
     "build": "craco build",
     "test": "craco test --testURL='http://localhost:7575'",
     "eject": "react-scripts eject",
@@ -68,5 +69,8 @@
     "react-scripts": "^4.0.2",
     "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.0",
     "typescript": "4.2.3"
+  },
+  "prettier": {
+    "semi": true
   }
 }
