+++ ui/src/components/MainLayout.tsx
@@ -2,40 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import React from "react";
-import MainView from "./MainView";
 import { useParty } from "@daml/react";
 import { Link, Route, useHistory } from "react-router-dom";
 import "@fontsource/alata";
 import { formatDate } from "./Common";
-
-type Props = {
-  onLogout: () => void;
-};
-
-type TabProps = { to: string; exact?: boolean; icon: string; label: string };
-
-const TabLink: React.FC<TabProps> = ({ to, label, icon, exact }) => {
-  return (
-    <Route
-      path={to}
-      exact={exact}
-      children={({ match }) => (
-        <Link
-          to={to}
-          className={
-            "flex flex-grow-0 h-9 items-center text-blue text-sm font-alata mr-3 ml-3 mt-1 mb-1 rounded" +
-            (match ? " tab-active" : " tab-hover")
-          }
-        >
-          <i
-            className={"ph-" + icon + " text-blueGray-400 text-2xl center m-4"}
-          />
-          {label}
-        </Link>
-      )}
-    />
-  );
-};
 
 const tabs = {
   profile: {
@@ -76,6 +46,31 @@ const tabs = {
   },
 };
 
+type TabProps = { to: string; exact?: boolean; icon: string; label: string };
+
+const TabLink: React.FC<TabProps> = ({ to, label, icon, exact }) => {
+  return (
+    <Route
+      path={to}
+      exact={exact}
+      children={({ match }) => (
+        <Link
+          to={to}
+          className={
+            "flex flex-grow-0 h-9 items-center text-blue text-sm font-alata mr-3 ml-3 mt-1 mb-1 rounded" +
+            (match ? " tab-active" : " tab-hover")
+          }
+        >
+          <i
+            className={"ph-" + icon + " text-blueGray-400 text-2xl center m-4"}
+          />
+          {label}
+        </Link>
+      )}
+    />
+  );
+};
+
 const sidebar: Map<string, Array<TabProps>> = new Map([
   ["Patient1", [tabs.profile, tabs.appointments, tabs.treatments, tabs.bills]],
   ["PrimaryCareProvider", [tabs.profile, tabs.referrals, tabs.patients]],
@@ -93,13 +88,15 @@ const sidebar: Map<string, Array<TabProps>> = new Map([
   ["InsuranceCompany", [tabs.claims]],
 ]);
 
-// Defines the page that is shown after logging in for a certain role.
-// If a role doesn't have a route here it defaults to '/' (the Profile page).
 const roleRoutes: Map<string, string> = new Map([
   ["InsuranceCompany", "/provider/claims"],
 ]);
 
-const MainScreen: React.FC<Props> = ({ onLogout }) => {
+type MainLayoutProps = {
+  onLogout: () => void;
+};
+
+const MainLayout: React.FC<MainLayoutProps> = ({ onLogout, children }) => {
   const [date] = React.useState(new Date());
   const role = useParty();
 
@@ -140,11 +137,9 @@ const MainScreen: React.FC<Props> = ({ onLogout }) => {
           Change Roles
         </button>
       </div>
-      <div className="relative bg-trueGray-100">
-        <MainView />
-      </div>
+      <div className="relative bg-trueGray-100">{children}</div>
     </div>
   );
 };
 
-export default MainScreen;
+export default MainLayout;
