+++ ui/src/index.tsx
@@ -5,6 +5,6 @@ import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
 import "phosphor-icons";
-import App from "./components/App";
+import App from "./App";
 
 ReactDOM.render(<App />, document.getElementById("root"));
