+++ ui/src/react-app-env.d.ts
@@ -1,6 +1,1 @@
-///
-/// Copyright (c) 2019, Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-/// SPDX-License-Identifier: Apache-2.0
-///
-
 /// <reference types="react-scripts" />
