+++ ui/src/views/appointments/Appointment.tsx
@@ -1,17 +1,13 @@
 import React from "react";
 import { useParams } from "react-router-dom";
 import { Main } from "@daml.js/healthcare-claims-processing";
-import { CreateEvent } from "@daml/ledger";
-import { useStreamQueries, useLedger } from "@daml/react";
 import { Clock } from "phosphor-react";
-import { mapIter, innerJoin, Message, useAsync, formatDate } from "./Common";
-import { ChoiceModal, FollowUp, creations } from "./ChoiceModal";
-import {
-  TabularScreenRoutes,
-  TabularView,
-  SingleItemView,
-} from "./TabularScreen";
+import { Message } from "components/Common";
+import { formatDate } from "utils";
+import { ChoiceModal, FollowUp, creations } from "components/ChoiceModal";
+import { SingleItemView } from "components/TabularScreen";
 import { Party, Time } from "@daml/types";
+import { useAppointments } from "hooks/appointments";
 
 type Props = {
   role: Party;
@@ -19,84 +15,6 @@ type Props = {
 
 const formatDateHelper = (timeStr: Time) =>
   timeStr ? formatDate(new Date(timeStr)) : "";
-
-const AppointmentRoutes: React.FC<Props> = ({ role }) => (
-  <TabularScreenRoutes
-    metavar=":appointmentId"
-    table={Appointments}
-    detail={Appointment({ role })}
-  />
-);
-
-const useAppointments = (query: any) => {
-  const ledger = useLedger();
-  const appointment = useAsync(
-    async () =>
-      query.appointmentId
-        ? await ledger.fetch(Main.Appointment.Appointment, query.appointmentId)
-        : null,
-    query
-  );
-  const appointmentsStream = useStreamQueries(
-    Main.Appointment.Appointment,
-    () => [query]
-  ).contracts;
-  const appointments: readonly CreateEvent<Main.Appointment.Appointment>[] =
-    query.appointmentId && appointment ? [appointment] : appointmentsStream;
-
-  const disclosed = useStreamQueries(Main.Policy.DisclosedPolicy).contracts;
-
-  const keyedAppointments = new Map(
-    appointments.map((p) => [p.payload.policy, p])
-  );
-  const keyedDisclosed = new Map(disclosed.map((p) => [p.contractId, p]));
-  return Array.from(
-    mapIter(
-      ([appointment, policy]) => ({ appointment, policy }),
-      innerJoin(keyedAppointments, keyedDisclosed).values()
-    )
-  );
-};
-
-const useAppointmentsData = () => useAppointments({});
-
-const Appointments: React.FC = () => {
-  return (
-    <TabularView
-      title="Appointments"
-      useData={useAppointmentsData}
-      fields={[
-        {
-          label: "Appointment Date",
-          getter: (o) =>
-            formatDateHelper(o?.appointment?.payload?.appointmentTime),
-        },
-        {
-          label: "Patient Name",
-          getter: (o) => o?.policy?.payload?.patientName,
-        },
-        {
-          label: "Insurance ID",
-          getter: (o) => o?.policy?.payload?.insuranceID,
-        },
-        {
-          label: "Procedure Code",
-          getter: (o) =>
-            o?.appointment?.payload?.encounterDetails.encounterDetails
-              .procedureCode,
-        },
-        {
-          label: "Appointment Priority",
-          getter: (o) =>
-            o?.appointment?.payload?.encounterDetails.encounterDetails
-              .appointmentPriority,
-        },
-      ]}
-      tableKey={(o) => o.appointment.contractId}
-      itemUrl={(o) => o.appointment.contractId}
-    />
-  );
-};
 
 const useAppointmentData = () => {
   const { appointmentId } = useParams<{ appointmentId: string }>();
@@ -175,7 +93,7 @@ const Appointment: React.FC<Props> = ({ role }) => {
       tableKey={(o) => o.overview?.appointment.contractId}
       itemUrl={(o) => ""}
       choices={(d) =>
-        d?.overview?.appointment?.payload?.provider === role ? (
+        d?.overview?.appointment?.payload?.provider === role && (
           <ChoiceModal
             className="flex flex-col space-y-6 w-170 mt-3"
             choice={Main.Appointment.Appointment.CheckInPatient}
@@ -208,12 +126,10 @@ const Appointment: React.FC<Props> = ({ role }) => {
               }
             />
           </ChoiceModal>
-        ) : (
-          <></>
         )
       }
     />
   );
 };
 
-export default AppointmentRoutes;
+export default Appointment;
