+++ ui/src/views/patients/Patient.tsx
@@ -7,89 +7,26 @@ import {
   useParams,
 } from "react-router-dom";
 import { Main } from "@daml.js/healthcare-claims-processing";
-import { CreateEvent } from "@daml/ledger";
 import { useParty, useStreamQueries } from "@daml/react";
 import { Share } from "phosphor-react";
 import {
-  mapIter,
-  innerJoin,
   FieldsRow,
   Message,
   PageTitleDiv,
   PageTitleSpan,
-  PageSubTitleSpan,
   TabLink,
-} from "./Common";
-import { useField } from "formik";
-import Select from "react-select";
-import {
-  LField,
-  EField,
-  ChoiceModal,
-  ChoiceErrorsType,
-  Nothing,
-  validateNonEmpty,
-  RenderError,
-} from "./ChoiceModal";
-import { TabularView } from "./TabularScreen";
+  PageSubTitleSpan,
+} from "components/Common";
+import { LField, EField, ChoiceModal, Nothing } from "components/ChoiceModal";
+import { usePatients } from "hooks/patients";
+import PolicySelectField from "components/PolicySelectField";
 
 type PatientOverview = {
   acceptance: Main.Patient.NotifyPatientOfPCPAcceptance;
   policy: Main.Policy.DisclosedPolicy;
 };
 
-const PatientRoutes: React.FC = () => {
-  const match = useRouteMatch();
-  return (
-    <Switch>
-      <Route path={`${match.path}/:patientId`}>
-        <Patient />
-      </Route>
-      <Route path={match.path}>
-        <Patients />
-      </Route>
-    </Switch>
-  );
-};
-
-const usePatients = (query: any, predicate: any = () => true) => {
-  const acceptances = useStreamQueries(
-    Main.Patient.NotifyPatientOfPCPAcceptance,
-    () => [query]
-  ).contracts.map((resp) => resp.payload);
-  const disclosedRaw = useStreamQueries(Main.Policy.DisclosedPolicy, () => [
-    query,
-  ]).contracts.filter((resp) => predicate(resp.payload));
-  const disclosed = disclosedRaw.map((resp) => resp.payload);
-
-  const keyedAcceptance = new Map(acceptances.map((p) => [p.patient, p]));
-  const keyedDisclosed = new Map(disclosed.map((p) => [p.patient, p]));
-  const overviews = Array.from(
-    mapIter(
-      ([acceptance, policy]) => ({ acceptance, policy }),
-      innerJoin(keyedAcceptance, keyedDisclosed).values()
-    )
-  );
-  return { acceptances, disclosed, overviews, disclosedRaw };
-};
-
-const useAllPatients: () => PatientOverview[] = () => usePatients({}).overviews;
-const Patients: React.FC = () => {
-  return (
-    <TabularView
-      title="Patients"
-      useData={useAllPatients}
-      fields={[
-        { label: "Name", getter: (o) => o.policy.patientName },
-        { label: "Insurance ID", getter: (o) => o.policy.insuranceID },
-      ]}
-      tableKey={(o) => o.policy.patient}
-      itemUrl={(o) => o.policy.patient}
-    />
-  );
-};
-
-const Patient: React.FC = () => {
+const SinglePatient: React.FC = () => {
   const username = useParty();
   const controlled = (d: Main.Policy.DisclosedPolicy) =>
     d.receivers.length > 0 && d.receivers.includes(username);
@@ -99,6 +36,7 @@ const Patient: React.FC = () => {
     { patient: patientId },
     controlled
   );
+
   const match = useRouteMatch();
 
   const policyRows = disclosed.map((d, i) => (
@@ -132,12 +70,10 @@ const Patient: React.FC = () => {
               buttonTitle="Refer Patient"
               icon={<Share />}
               successWidget={({ rv: [v, evts] }, close) => (
-                <>
-                  <Message
-                    title="Referral Created!"
-                    content="Change to the Radiologist role to see the referral and schedule an appointment with the patient."
-                  />
-                </>
+                <Message
+                  title="Referral Created!"
+                  content="Change to the Radiologist role to see the referral and schedule an appointment with the patient."
+                />
               )}
               initialValues={{
                 policy: Nothing,
@@ -152,7 +88,7 @@ const Patient: React.FC = () => {
               {({ errors, touched }) => (
                 <>
                   <h1 className="heading-2xl mb-7">Create Referral</h1>
-                  <PolicySelect
+                  <PolicySelectField
                     label="Policy"
                     name="policy"
                     disclosedRaw={disclosedRaw}
@@ -198,9 +134,12 @@ const Patient: React.FC = () => {
           <hr />
           <FieldsRow
             fields={[
-              { label: "Name", value: po.policy.patientName },
-              { label: "Insurance ID", value: po.policy.insuranceID },
-              { label: "Primary Care Provider", value: "" },
+              { label: "Name", value: po?.policy?.patientName },
+              { label: "Insurance ID", value: po?.policy?.insuranceID },
+              {
+                label: "Primary Care Provider",
+                value: po?.acceptance?.provider,
+              },
             ]}
           />
         </Route>
@@ -230,45 +169,4 @@ const Patient: React.FC = () => {
   );
 };
 
-const PolicySelect: React.FC<{
-  name: string;
-  label: string;
-  disclosedRaw: readonly CreateEvent<Main.Policy.DisclosedPolicy>[];
-  errors?: ChoiceErrorsType;
-}> = ({ name, label, disclosedRaw, errors }) => {
-  const [, , helpers] = useField({
-    name,
-    validate: validateNonEmpty(label),
-  });
-  const { setValue } = helpers;
-  const formatOptionLabel = (a: CreateEvent<Main.Policy.DisclosedPolicy>) => (
-    <div className="">
-      Policy Provider: <b>{a.payload.payer}</b>
-      <br />
-      Disclosed Parties: <b>{a.payload.receivers}</b>
-      <br />
-      <div className="overflow-ellipsis-20">
-        Contract ID: <b>{a.contractId}</b>
-      </div>
-    </div>
-  );
-  const error = errors?.[name];
-  return (
-    <div className="flow flow-col mb-2 mt-0.5">
-      <label htmlFor={name} className="block label-sm">
-        {label}
-      </label>
-      <Select
-        classNamePrefix="react-select-modal-enum"
-        options={disclosedRaw}
-        onChange={(option) => setValue(option?.contractId)}
-        formatOptionLabel={formatOptionLabel}
-        getOptionValue={(a) => a.contractId}
-        styles={{ singleValue: (base) => ({ textOverflow: "ellipsis" }) }}
-      />
-      <RenderError error={error} />
-    </div>
-  );
-};
-
-export default PatientRoutes;
+export default SinglePatient;
