+++ ui/src/views/patients/Patients.tsx
@@ -1,0 +1,26 @@
+import { Main } from "@daml.js/healthcare-claims-processing";
+import { TabularView } from "components/TabularScreen";
+import { usePatients } from "hooks/patients";
+
+type PatientOverview = {
+  acceptance: Main.Patient.NotifyPatientOfPCPAcceptance;
+  policy: Main.Policy.DisclosedPolicy;
+};
+
+const useAllPatients: () => PatientOverview[] = () => usePatients({}).overviews;
+const TablePatients: React.FC = () => {
+  return (
+    <TabularView
+      title="Patients"
+      useData={useAllPatients}
+      fields={[
+        { label: "Name", getter: (o) => o?.policy?.patientName },
+        { label: "Insurance ID", getter: (o) => o?.policy?.insuranceID },
+      ]}
+      tableKey={(o) => o?.policy?.patient}
+      itemUrl={(o) => o?.policy?.patient}
+    />
+  );
+};
+
+export default TablePatients;
